pyproject.toml
[tool.poetry]
name = "drf-serializers-and-filters-24-2"
version = "0.1.0"
description = ""
authors = ["Gennady_Gubin"]
readme = "README.md"
packages = [{include = "drf_serializers_and_filters_24"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
ipython = "^8.18.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
